# WebAPI/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Web API ",
      "description": "API del sistema de capacitación",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
        "/api/getMision/{idMision}": {
            "get": {
                "tags": [
                    "Mision"
                ],
                "summary": "Encuentra una misión por id",
                "description": "Regresa la misión con el id solicitado",
                "operationId": "getMision",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a consultar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }
                    }
                }
            }
        },
        "/api/addMision": {
            "post": {
                "tags": [
                    "Mision"
                ],
                "summary": "Agrega una nueva misión",
                "description": "Agrega una misión a la base de datos",
                "operationId": "addMision",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nueva misión",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateMision/{idMision}": {
            "put": {
                "tags": [
                    "Mision"
                ],
                "summary": "Actualiza una misión",
                "description": "Actualiza una misión en la base de datos",
                "operationId": "updateMision",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos de una misión",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }
                    },
                    "400": {
                        "description": "Misión Inválida"
                    },
                    "404": {
                        "description": "Misión no encontrada"
                    }

                }
            }
        },
        "/api/deleteMision/{idMision}": {
            "delete": {
                "tags": [
                    "Mision"
                ],
                "summary": "Borra una misión",
                "description": "Inactiva una misión de la base de datos",
                "operationId": "deleteMision",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                      "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a borrar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/Mision"
                        }
                    },
                    "400": {
                        "description": "Misión Inválida"
                    },
                    "404": {
                        "description": "Misión no encontrada"
                    }
                }
            }
        },
...
        "/api/getMisionesEmpleado/{idEmpleado}": {
            "get": {
                "tags": [
                    "MisionEmpleado"
                ],
                "summary": "Encuentra todas las misiones superadas por un empleado, mediante su id",
                "description": "Regresa todas las misiones superadas por un empleado, mediante la solicitud de su id",
                "operationId": "getMisionesEmpleado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }
                    }
                }
            }
        },
        "/api/getMisionEmpleado/{idMision}/{idEmpleado}": {
            "get": {
                "tags": [
                    "MisionEmpleado"
                ],
                "summary": "Encuentra una misión superada por un empleado, mediante el id de ambos",
                "description": "Regresa la misión superada por un empleado, mediante el id solicitado de ambos",
                "operationId": "getMisionEmpleado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a obtener",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }
                    }
                }
            }
        },
        "/api/addMisionEmpleado": {
            "post": {
                "tags": [
                    "MisionEmpleado"
                ],
                "summary": "Agrega a un empleado una misión superada por él",
                "description": "Agrega a un empleado una misión superada por él a la base de datos",
                "operationId": "addMisionEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Nueva misión superada por un empleado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }
                    },
                    "405": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/updateMisionEmplado/{idMision}/{idEmpleado}": {
            "put": {
                "tags": [
                    "MisionEmpleado"
                ],
                "summary": "Actualiza el estatus de realización de una misión a un empleado",
                "description": "Actualiza el estatus de realización de una misión a un empleado en la base de datos",
                "operationId": "updateMisionEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos nuevos de la misión superada de un empleado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }                      
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }
                    },
                    "400": {
                        "description": "Mision de empleado Inválida"
                    },
                    "404": {
                        "description": "Mision de empleado no encontrada"
                    }

                }
            }
        },
        "/api/deleteMisionEmpleado/{idMision}/{idEmpleado}": {
            "delete": {
                "tags": [
                    "MisionEmpleado"
                ],
                "summary": "Borra la misión de empleado",
                "description": "Borra la misión de un empleado de la base de datos",
                "operationId": "deleteMisionEmpleado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idMision",
                        "description": "Id de la misión a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "idEmpleado",
                        "description": "Id del empleado a actualizar",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MisionEmpleado"
                        }
                    },
                    "400": {
                        "description": "Mision de empleado Inválida"
                    },
                    "404": {
                        "description": "Mision de empleado no encontrada"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
       "Mision": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "habilidad": {
                    "type": "string"
                },
                "dificultad": {
                    "type": "integer",
                    "format": "int64"
                },
                "puntos": {
                    "type": "integer",
                    "format": "int64"
                },
                "numero": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "MisionEmpleado": {
            "type": "object",
            "properties": {
                "idMision": {
                    "type": "integer",
                    "format": "int64"
                },
                "idEmpleado": {
                    "type": "integer",
                    "format": "int64"
                },
                "fechaCurso": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Fecha y hora en que el empleado cursó la misión",
                    "example": "2021-07-21T17:32:28Z"
                }
            }
        },
...
    }
  }